requirements.txt
aiosqlite==0.17.0
anyio==3.3.3
asgiref==3.4.1
asyncpg==0.24.0
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.2
iso8601==0.1.16
passlib==1.7.4
pycodestyle==2.8.0
pydantic==1.8.2
PyJWT==2.3.0
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
